throws error if user attempts to update their email with the value that it already is
hot fix for https://github.com/apatrickstar/AllMyPPL-Client-iOS/issues/4

Somebody asked a question about this on Parse.com questions, and the
devs confirmed that email verification is required every time the value
of that field is updated on a user, even if its the already existing
value, its a way to manually trigger another verification.  In our
case, we only want to both verifying a user’s email if its never been,
so we throw an error when attempting to save the email field with the
value it has already.
https://parse.com/questions/verify-email-resend-confirmation-email